requirement.txt
gensim==3.5.0
lightgbm==2.1.2
matplotlib==2.2.2
mxnet-cu90==1.2.1.post1
numpy==1.14.5
pandas==0.23.2
scikit-learn==0.19.2
scipy==1.1.0
